rename cellfinder to cellfinder_brainreg
