+++ brainglobe_workflows/cellfinder_brainreg/analyse/analyse.py
@@ -18,7 +18,9 @@ from bg_atlasapi import BrainGlobeAtlas
 from brainglobe_utils.general.system import ensure_directory_exists
 from brainglobe_utils.pandas.misc import safe_pandas_concat, sanitise_df
 
-from brainglobe_workflows.export.export import export_points
+from brainglobe_workflows.cellfinder_brainreg.export.export import (
+    export_points,
+)
 
 
 class Point:
