+++ brainglobe_workflows/cellfinder_brainreg/extract/extract_cubes.py
@@ -22,8 +22,10 @@ from skimage import transform
 from tifffile import tifffile
 from tqdm import tqdm
 
-from brainglobe_workflows.tools import image_processing as img_tools
-from brainglobe_workflows.tools import system
+from brainglobe_workflows.cellfinder_brainreg.tools import (
+    image_processing as img_tools,
+)
+from brainglobe_workflows.cellfinder_brainreg.tools import system
 
 
 class StackSizeError(Exception):
