+++ tests/cellfinder_brainreg/test_integration/test_registration.py
@@ -7,7 +7,9 @@ import pandas as pd
 import pytest
 from imio.load import load_any
 
-from brainglobe_workflows.main import main as cellfinder_run
+from brainglobe_workflows.cellfinder_brainreg.main import (
+    main as cellfinder_run,
+)
 
 data_dir = os.path.join(
     os.getcwd(),
