+++ tests/cellfinder_brainreg/test_unit/test_tools/test_image_processing.py
@@ -2,7 +2,9 @@ import random
 
 import numpy as np
 
-from brainglobe_workflows.tools import image_processing as img_tools
+from brainglobe_workflows.cellfinder_brainreg.tools import (
+    image_processing as img_tools,
+)
 
 
 def test_crop_center_2d():
