+++ tests/cellfinder_brainreg/test_unit/test_tools/test_prep.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from brainglobe_utils.general.exceptions import CommandLineInputError
 
-from brainglobe_workflows.tools import prep
+from brainglobe_workflows.cellfinder_brainreg.tools import prep
 
 # import shutil
