+++ tests/cellfinder_brainreg/test_unit/test_tools/test_system.py
@@ -6,9 +6,9 @@ import pytest
 from brainglobe_utils.general.exceptions import CommandLineInputError
 from brainglobe_utils.general.system import ensure_directory_exists
 
-import brainglobe_workflows.tools.system as system
+import brainglobe_workflows.cellfinder_brainreg.tools.system as system
 
-data_dir = Path("tests", "data")
+data_dir = Path(__file__).parents[4] / Path("tests", "data")
 background_im_dir = os.path.join(data_dir, "background")
